Move Exceptions to own namespace
