+++ wcfsetup/install/files/lib/system/message/unfurl/UnfurlResponse.class.php
@@ -11,6 +11,9 @@ use GuzzleHttp\Psr7\Response;
 use GuzzleHttp\RequestOptions;
 use Psr\Http\Client\ClientExceptionInterface;
 use wcf\system\io\HttpFactory;
+use wcf\system\message\unfurl\exception\DownloadFailed;
+use wcf\system\message\unfurl\exception\ParsingFailed;
+use wcf\system\message\unfurl\exception\UrlInaccessible;
 use wcf\util\ArrayUtil;
 use wcf\util\StringUtil;
 use wcf\util\Url;
